requirements.txt
torch --index-url https://download.pytorch.org/whl/cu118
deepspeed
einops
evaluate
ijson
ipdb
jieba
joblib
nltk
numpy
openai
pandas
peft
protobuf
PyPDF2
PyYAML
redis
Requests
rouge_score
scipy
torch
tqdm
transformers
loguru
tensorboard
xmltodict
flask
sentencepiece
bitsandbytes
bert_score
mauve-text